# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rokwire Groups Building Block API Documentation.",
        "title": "Rokwire Groups Building Block API",
        "contact": {},
        "version": "1.29.0"
    },
    "host": "localhost",
    "basePath": "/gr",
    "paths": {
...
        "/api/analytics/groups": {
            "get": {
                "security": [
                    {
                        "IntAPIKeyAuth": []
                    }
                ],
                "description": "Gets groups",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "operationId": "AnalyticsGetGroups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date string - RFC3339 encoded",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date string - RFC3339 encoded",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.analyticsGetGroupsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/members": {
            "get": {
                "security": [
                    {
                        "IntAPIKeyAuth": []
                    }
                ],
                "description": "Gets groups members",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "operationId": "AnalyticsGetGroupsMembers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date string - RFC3339 encoded",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date string - RFC3339 encoded",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.analyticsGetGroupsMembersResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/posts": {
            "get": {
                "security": [
                    {
                        "IntAPIKeyAuth": []
                    }
                ],
                "description": "Gets posts",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "operationId": "AnalyticsGetPosts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date string - RFC3339 encoded",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date string - RFC3339 encoded",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.analyticsGetPostsResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "GroupStats": {
            "type": "object",
            "properties": {
                "admins_count": {
                    "type": "integer"
                },
                "attendance_count": {
                    "type": "integer"
                },
                "member_count": {
                    "type": "integer"
                },
                "pending_count": {
                    "type": "integer"
                },
                "rejected_count": {
                    "type": "integer"
                },
                "total_count": {
                    "description": "pending and rejected are excluded",
                    "type": "integer"
                }
            }
        },
...
        "rest.analyticsGetGroupsMembersResponse": {
            "type": "object",
            "properties": {
                "org_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_updated": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "rest.analyticsGetGroupsResponse": {
            "type": "object",
            "properties": {
                "authman_enabled": {
                    "type": "boolean"
                },
                "authman_group": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "org_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_updated": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hidden_for_search": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "privacy": {
                    "type": "string"
                },
                "research_group": {
                    "type": "boolean"
                },
                "research_open": {
                    "type": "boolean"
                },
                "stats": {
                    "$ref": "#/definitions/GroupStats"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "rest.analyticsGetPostsResponse": {
            "type": "object",
            "properties": {
                "org_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_updated": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "member_user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "ROKWIRE-API-KEY",
            "in": "header"
        },
        "AppUserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add Bearer prefix to the Authorization value)"
        },
        "IntAPIKeyAuth": {
            "type": "apiKey",
            "name": "INTERNAL-API-KEY",
            "in": "header"
        }
    }
}